Extract strategy-briefing helpers into a partial
Increment 2d of the endpoint-monolith decomposition. Moves the strategy run
briefing helper cluster out of WorkstationEndpoints.cs into a new sibling partial
WorkstationEndpoints.StrategyBriefing.cs.

Moved (whole private static methods): BuildRunDigest, BuildStrategyRunCard,
BuildBriefingInsightFeed, BuildBriefingRun, BuildSavedComparisons, BuildBriefingAlerts,
BuildWhatChangedItems, BuildInsightSummary, BuildComparisonSummary, BuildChangeSummary,
GetInsightTone, GetSecurityCoverageIssueCount, BuildStrategyDrillInLinks, BuildTimelineCard,
BuildModeComparisons, BuildRunDrillInLinks, RunRoute, and both ParseModes overloads.

Left in core: ResolveModeVariant (a shared strategy-mode helper also used by the PlotTool
partial) and BuildAccountingRunCard (accounting). Because WorkstationEndpoints is a single
partial class this is a behavior-preserving relocation: core-resident callers still reach
BuildRunDigest / BuildStrategyRunCard across the partial with no signature/route/DI change.
Core drops 7248 -> 6865 lines; ratchet baseline retightened.
